gemini/requirements.txt
Flask
gunicorn
numpy
IPython
python-dotenv